-Initial attempt at constraining/standardizing the types used. Limit to long, double, string and IntPtr. -Fix NumPut() and NumGet() to better handle objects with ptr and size properties. -Fix bug where RegRead() was returning an array of bytes, when it should be an array of longs so the element type is compatible with the rest of the program. -Fix and disable some warnings. -Use more collection initialization expressions. -Misc cleanup.
https://bitbucket.org/mfeemster/keysharp/issues/240/standardize-data-types

+++ Keysharp.Core/Windows/WindowsHookThread.cs
@@ -5074,7 +5074,7 @@ namespace Keysharp.Core.Windows
 
 									break;
 
-								case (uint)WM_HOTKEY://Some hotkeys are handled directly by windows using WndProc(), others, such as those with left/right modifiers, are handled directly by us.
+								case WM_HOTKEY://Some hotkeys are handled directly by windows using WndProc(), others, such as those with left/right modifiers, are handled directly by us.
 								case (uint)UserMessages.AHK_HOOK_HOTKEY://Some hotkeys are handled directly by windows using WndProc(), others, such as those with left/right modifiers, are handled directly by us.
 								{
 									Script.HookThread.kbdMsSender.ProcessHotkey((int)wParamVal, (int)lParamVal, msg.obj as HotkeyVariant, msg.message);
